+++ src/export_roms/test_iso_to_rvz.rs
@@ -77,6 +77,7 @@ async fn test() {
         &RvzCompressionAlgorithm::Zstd,
         5,
         128,
+        false,
     )
     .await
     .unwrap();