+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Change `-s` in various subcommands to accept SQL wildcards
 - Change `-n` to `-g` (for game) in various subcommands and make it accept SQL wildcards
 - Change `-g` to `-o` (for one g one r) in various subcommands
+- Add an option to scrub RVZ in `export-roms`
 
 # 0.18.1
 - Bump minimum required chdman version to 0.264 for dreamcast